Fix migrations to work with Sqlite, Azure SQL and SQL Server Changed price from double to decimal

+++ ThAmCo.Orders.Api/Data/OrderDetail.cs
@@ -3,7 +3,7 @@
         public int Id { get; set; }
         public int OrderId { get; set; }
         public int ProductId { get; set; }
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public int Quantity { get; set; }
         public Product Product { get; set; } = null!;
     }
